requirements.txt
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.13.0
annotated-types==0.6.0
anyio==4.1.0
apache-airflow==2.8.0
apache-airflow-providers-common-sql==1.9.0
apache-airflow-providers-ftp==3.7.0
apache-airflow-providers-http==4.8.0
apache-airflow-providers-imap==3.5.0
apache-airflow-providers-sqlite==3.6.0
apispec==6.3.0
apturl==0.5.2
argcomplete==3.2.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
Babel==2.14.0
backoff==2.2.1
bcrypt==3.2.0
blinker==1.7.0
boto3==1.34.29
botocore==1.34.29
Brlapi==0.8.3
Brotli==1.0.9
cachelib==0.9.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
command-not-found==0.3
configobj==5.0.8
configparser==6.0.0
ConfigUpdater==3.2
confluent-kafka==2.3.0
connexion==2.14.2
cron-descriptor==1.4.0
croniter==2.0.1
cryptography==41.0.7
cssselect==1.2.0
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
Deprecated==1.2.14
dill==0.3.1.1
discord==2.3.2
discord.py==2.3.2
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0
djangorestframework==3.14.0
dnspython==2.4.2
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.20.1
duplicity==0.8.21
email-validator==1.3.1
etelemetry==0.3.1
exceptiongroup==1.2.0
fasteners==0.14.1
filelock==3.6.0
findspark==2.0.1
fitz==0.0.1.dev2
Flask==2.2.5
Flask-AppBuilder==4.3.10
Flask-Babel==2.0.0
Flask-Caching==2.1.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
frontend==0.0.3
frozenlist==1.4.1
fsspec==2023.12.2
future==0.18.2
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.26.1
google-generativeai==0.3.2
google-re2==1.1
googleapis-common-protos==1.62.0
gpg==1.16.0
graphviz==0.20.1
greenlet==3.0.2
grpcio==1.60.0
grpcio-status==1.60.0
gunicorn==21.2.0
gyp==0.1
h11==0.14.0
halo==0.0.31
httpcore==0.16.3
httplib2==0.20.2
httpx==0.23.3
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
inflection==0.5.1
isodate==0.6.1
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.10.0
libcomps==0.1.15
limits==3.7.0
linkify-it-py==2.0.2
lockfile==0.12.2
log-symbols==0.0.14
looseversion==1.3.0
louis==3.20.0
lxml==5.0.0
macaroonbakery==1.3.1
Mako==1.3.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.4.0
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.4
mutagen==1.45.1
netifaces==0.11.0
networkx==3.2.1
nibabel==5.2.0
nipype==1.8.6
numpy==1.26.2
oauthlib==3.2.0
olefile==0.46
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
ordered-set==4.1.0
packaging==23.2
pandas==2.2.0
paramiko==3.4.0
parquet-tools==0.2.15
pathlib==1.0.1
pathspec==0.12.1
pdf2image==1.16.3
pdfminer.six==20231228
pdfquery==0.4.3
pendulum==2.1.2
pexpect==4.8.0
Pillow==9.0.1
pipreqs==0.4.13
platformdirs==2.5.1
pluggy==1.3.0
prison==0.2.1
proto-plus==1.23.0
protobuf==4.25.1
prov==2.0.0
psutil==5.9.6
psycopg2==2.9.9
ptyprocess==0.7.0
py4j==0.10.9.7
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycairo==1.20.1
pycparser==2.21
pycryptodomex==3.11.0
pycups==2.0.1
pydantic==2.5.2
pydantic_core==2.14.5
pydot==2.0.0
Pygments==2.17.2
PyGObject==3.42.1
PyHive==0.7.0
PyJWT==2.8.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==3.1.1
pypdf==3.17.4
PyPDF2==3.0.1
pyquery==2.0.0
pyRFC3339==1.1
pyrsistent==0.18.1
pyspark==3.5.0
pytesseract==0.3.10
python-apt==2.4.0+ubuntu2
python-daemon==3.0.1
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2023.3.post1
pytzdata==2020.1
pyxattr==0.7.2
pyxdg==0.27
pyxnat==1.6
PyYAML==6.0.1
rdflib==7.0.0
referencing==0.32.0
reportlab==3.6.8
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==13.7.0
rich-argparse==1.4.0
roman==4.1
rpds-py==0.13.2
rpm==4.17.0
rsa==4.9
s3transfer==0.10.0
scipy==1.11.4
SecretStorage==3.3.1
setproctitle==1.3.3
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
spinners==0.0.24
SQLAlchemy==1.4.50
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
ssh-import-id==5.11
starlette==0.34.0
systemd-python==234
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
text-unidecode==1.3
texttable==1.6.4
thrift==0.16.0
tqdm==4.66.1
traits==6.3.2
typing_extensions==4.9.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
uc-micro-py==1.0.2
ufw==0.36.1
unattended-upgrades==0.1
unicodecsv==0.14.1
universal-pathlib==0.1.4
urllib3==2.0.6
usb-creator==0.3.7
uvicorn==0.25.0
virtualenv==20.13.0+ds
wadllib==1.3.6
websocket-client==1.2.3
websockets==9.1
Werkzeug==2.2.3
wrapt==1.16.0
WTForms==3.1.1
xdg==5
xkit==0.0.0
yarg==0.1.9
yarl==1.9.4
yt-dlp==2022.4.8
zipp==3.17.0
